backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
boto3==1.34.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
mediapipe==0.10.11
opencv-python==4.9.0.80
